+++ Sources/RepoBarCore/Support/AuthDefaults.swift
@@ -6,4 +6,5 @@ public enum RepoBarAuthDefaults {
     public static let loopbackPort: Int = 53682
     public static let githubHost = URL(string: "https://github.com")!
     public static let apiHost = URL(string: "https://api.github.com")!
+    public static let appInstallURL = URL(string: "https://github.com/apps/repobar/installations/new")!
 }